Move Elasticsearch-specific methods to the appropriate place and remove obsolete ones
